Time: 21 ms (97.53%), Space: 77.6 MB (74.74%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.java
@@ -1,14 +1,14 @@
 class Solution {
     public int maximumCandies(int[] candies, long k) {
-        long l=0;
+        int l=0;
         long sum=0;
         for(int i:candies)
             sum+=i;
-        long r=(sum/k);
+        int r=(int)(sum/k);
         
-        long ans=0;  // Initially our answer is this
+        int ans=0;  // Initially our answer is this
         while(l<=r){
-            long m=l+(r-l)/2;
+            int m=l+(r-l)/2;
             long dis=k;
             if(m!=0){
                 dis=0;
@@ -22,6 +22,6 @@ class Solution {
             }
             else r=m-1;
         }
-        return (int)ans;
+        return ans;
     }
 }
